feat(crypto): 코인 거래 인터페이스 추상화 + Market Profile CRYPTO scope 추가
- BrokerClient / MarketDataProvider / RealtimeProvider Protocol 정의 (trading/broker_base.py)
- MarketScannerProtocol 정의 (agent/scanner_base.py)
- Market enum에 BITHUMB 추가, MARKET_SCOPE_CRYPTO 상수 + is_crypto_market() 헬퍼
- normalize_market_scope()에 CRYPTO 분기, markets_for_scope() CRYPTO 지원
- OrderRequest/OrderResult/HoldingInfo quantity int→float (코인 소수점 수량)

+++ trading/models.py
@@ -37,7 +37,7 @@ class CurrentPrice(BaseModel):
     exchange_rate_to_krw: float = 1.0
     change: float
     change_rate: float
-    volume: int
+    volume: float
     session: str = ""
     timestamp: datetime
 
@@ -48,7 +48,7 @@ class OrderRequest(BaseModel):
     market: Market
     side: OrderSide
     order_type: OrderType
-    quantity: int
+    quantity: float
     price: Optional[float] = None  # 시장가 주문 시 None
     currency: str = "KRW"
     exchange_rate_to_krw: float = 1.0
@@ -59,7 +59,7 @@ class OrderResult(BaseModel):
     success: bool
     order_id: Optional[str] = None
     message: str
-    filled_quantity: int = 0
+    filled_quantity: float = 0
     filled_price: float = 0.0
     currency: str = "KRW"
     filled_price_krw: float = 0.0
@@ -91,7 +91,7 @@ class HoldingInfo(BaseModel):
     name: str
     market: str = "KRX"
     currency: str = "KRW"
-    quantity: int
+    quantity: float
     avg_buy_price: float
     current_price: float
     pnl: float
@@ -107,9 +107,9 @@ class PendingOrderInfo(BaseModel):
     market: str = "KRX"
     currency: str = "KRW"
     side: str               # 매수/매도
-    order_qty: int          # 주문수량
-    filled_qty: int         # 체결수량
-    remaining_qty: int      # 미체결수량
+    order_qty: float        # 주문수량
+    filled_qty: float       # 체결수량
+    remaining_qty: float    # 미체결수량
     order_price: float      # 주문단가
     order_time: str         # 주문시각
     exchange_rate_to_krw: float = 1.0
